chore: rename CreatorList component
